renamed winston/logging config
